==> Code/Workflow/9 - Results reported 2024-07-19.R <==
#######################################
#######################################
## Author: Dr. Graziella DiRenzo
##
## Date Created: 2022-05-12
##
## Copyright (c) Graziella DiRenzo, 2022
## Email: gdirenzo@umass.edu
#######################################
#######################################


#######################################
## Code objectives:
#######################################



# To generate the values reported in the results section of the Globi manuscript



#######################################
## Input Data:
#######################################


# Model output files (from Script 5):
#   - out-bee_species-with-priors-1-NIMBLE.rds
#   - result-bee_species-with-priors-1-NIMBLE.rds

# Data info (from Script 3):
#   - dat_info_2025_01_22.rds



#######################################
## Output of code:
#######################################


# None - this script calculates and prints the statistics reported in the
# results section of the manuscript (no files are saved)



#######################################
############ Table of Contents ########
#######################################


# 1. Load libraries & set working directory
# 2. Load data
# 3. Calculate values reported in the results

#######################################
#######################################
#######################################




# 1. Load libraries & set working directory -------------------------------------------------------



# Load libraries
library(tidyverse)
library(mcmcr)


# Add github path
github_path <- ""

# Add globi folder name
globi_out_folder    <- ""
globi_result_folder <- ""


# date object for folder
date_folder <- "2025 08 05"







# 2. Load data -------------------------------------------------------






#---- bee_species model

# Model name
mod_name <- "bee_species"

# Load the model output - out
load(paste0(globi_out_folder, "/out-bee_species-with-priors-1-NIMBLE.rds"))

# Save the out object with a new model specific name
out_bee_species <- out

out_bee_species_df <- as.data.frame(out_bee_species)


# Load the model output - result
load(paste0(globi_result_folder, "/result-quarter-bee_species-with-priors-1-NIMBLE.rds"))

# Save the out object with a new model specific name
result_bee_species <- subset_result_quarter




#------------  Upload the data & format
# object name = bee.plant.cite
# 3-D array
load("/Users/gdirenzo/Documents/GitHub/globi_msomAndsciChecklists/Data/data_summary/globi_data_formatted_bee_plant_date_citation_2025_01_22 - short plant list - no apis.rds")



# Read in the dat_info
load("/Users/gdirenzo/Documents/GitHub/globi_msomAndsciChecklists/Data/dat_info_2025_01_22.rds")
# object name = dat_info



# Load covariates
load("/Users/gdirenzo/Documents/GitHub/globi_msomAndsciChecklists/Data/model_covariates - 2025 01 22 - no apis.rds")



# Load in the number of observations
load("/Users/gdirenzo/Documents/GitHub/globi_msomAndsciChecklists/Data/obs_dat-2025-02-11.rds")
# object name = obs_dat










# 3. Calculate values reported in the results -------------------------------------------------------

# number of bees
nrow(covariates$bee.covariates)


## Calcualte number of solitary bees
sum(covariates$bee.covariates$solitary)
sum(covariates$bee.covariates$notSolitary)


## Calcualte number of striped bees
sum(covariates$bee.covariates$striped)
sum(covariates$bee.covariates$notStriped)

# Mean bee size
mean(covariates$bee.covariates$size)
min(covariates$bee.covariates$size)
max(covariates$bee.covariates$size)

# Number of plants
nrow(covariates$plant.covariates)


## Plant flower color sample size
sum(covariates$plant.covariates$yellow)
sum(covariates$plant.covariates$blue.new)
sum(covariates$plant.covariates$white)
sum(covariates$plant.covariates$other)


# Bowl vs not bowl
sum(covariates$plant.covariates$bowl)
# not bowl
nrow(covariates$plant.covariates) - sum(covariates$plant.covariates$bowl)


# Source types
sum(covariates$citation.covariates$observation)
sum(covariates$citation.covariates$collection)
sum(covariates$citation.covariates$literature)
sum(covariates$citation.covariates$aggregated)


##### Calculate: Number of total unique bee-plant interactions in raw data

# Collapse across date & citation
y.bee.plant <- apply(bee.plant.cite, c(1, 2), max, na.rm = TRUE)
y.bee.plant[y.bee.plant == "-Inf"] <- 0

# Total
sum(y.bee.plant)


##### Calculate: Max number of interactions per bee species
y.bee.plant <- apply(y.bee.plant, 1, sum, na.rm = TRUE)

max(y.bee.plant)

y.bee.plant[which(y.bee.plant == max(y.bee.plant))]


##### Calculate: Mean number of interactions per bee species
mean(y.bee.plant)

sd(y.bee.plant)/sqrt(length(y.bee.plant))

min(y.bee.plant)

max(y.bee.plant)

##### Calculate: Number of bee species with 0 plant interactions

length(which(y.bee.plant == 0))


##### Calculate: Number of bee species with ==1 plant interactions

length(which(y.bee.plant ==1))

##### Calculate: Number of bee species with ==2 plant interactions

length(which(y.bee.plant ==2))

##### Calculate: Number of bee species with ==3 plant interactions

length(which(y.bee.plant ==3))


##### Calculate: Number of bee-plant interactions from occupancy model



# Summarize z
# # Latent state variables
out_num <- as.mcmc(data.frame(rbind(result_bee_species[[1]]$samples2[,grep("n_plants_per_bee", colnames(result_bee_species[[1]]$samples2))],
                                    result_bee_species[[2]]$samples2[,grep("n_plants_per_bee", colnames(result_bee_species[[1]]$samples2))],
                                    result_bee_species[[3]]$samples2[,grep("n_plants_per_bee", colnames(result_bee_species[[1]]$samples2))])))



# Extract the mean values for z = true bee, plant, by month interactions
# Look at the number of dimensions
dim(out_num)


# Now, we want the number of unique plants per bee

# To do this, we first calculate the mean number of interactions per bee species by taking the column mean
plant_interactions_per_bee <- apply(out_num, 2, mean)


mean(plant_interactions_per_bee)

sd(plant_interactions_per_bee)/sqrt(length(plant_interactions_per_bee))

min(plant_interactions_per_bee)

max(plant_interactions_per_bee)








# End script
